chore: use flat route
