+++ .tnfrc.ts
@@ -1,5 +1,8 @@
 export default {
   router: {
     defaultPreload: 'intent',
+    convention: {
+      routeFileIgnorePattern: 'components',
+    },
   },
 };
