+++ src/components/Account/Applications/index.style.ts
