--- src/pages/list/_search/search/projects/-service.ts
@@ -1,10 +1,0 @@
-import { request } from '@/utils/request';
-import type { ListItemDataType, Params } from './-data';
-
-export async function queryFakeList(
-  params: Params,
-): Promise<{ data: { list: ListItemDataType[] } }> {
-  return request('/api/fake_list', {
-    params,
-  });
-}
