--- src/pages/list/card-list/-service.ts
@@ -1,10 +1,0 @@
-import { request } from '@/utils/request';
-import type { CardListItemDataType } from './-data.d';
-
-export async function queryFakeList(params: {
-  count: number;
-}): Promise<{ data: { list: CardListItemDataType[] } }> {
-  return request('/api/card_fake_list', {
-    params,
-  });
-}
