+++ src/styles/search-applications.ts
