+++ src/styles/search-articles.ts
