+++ src/styles/search-projects.ts
